pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "modelguard"
version = "0.1.0"
description = "Data Drift, Model Decay & Auto-Retraining System"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "ModelGuard Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    "sqlalchemy>=2.0",
    "alembic>=1.12",
    "fastapi>=0.104",
    "uvicorn>=0.24",
    "pydantic>=2.5",
    "typer>=0.9",
    "rich>=13.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "structlog>=23.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "mypy>=1.7",
    "pre-commit>=3.6",
]
notifications = [
    "slack-sdk>=3.23",
    "sendgrid>=6.10",
]
mlflow = [
    "mlflow>=2.9",
]
all = [
    "modelguard[dev,notifications,mlflow]",
]

[project.scripts]
modelguard = "modelguard.cli.main:app"

[project.urls]
Homepage = "https://github.com/Aagam-Bothara/ModelGuard"
Repository = "https://github.com/Aagam-Bothara/ModelGuard"

[tool.hatch.build.targets.wheel]
packages = ["src/modelguard"]

[tool.ruff]
target-version = "py310"
line-length = 88
select = ["E", "F", "I", "N", "W", "B", "C4", "UP"]
ignore = ["E501"]

[tool.ruff.isort]
known-first-party = ["modelguard"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=modelguard --cov-report=term-missing"
